+++ docs/release-notes.md
@@ -6,6 +6,7 @@
   * Improved performance.
   * Improved error message when a mod is blocked by Windows Smart App Control.
   * Improved translations. Thanks to To2morrow (updated Korean)!
+  * Updated internal mod blacklist.
 
 * For mod authors:
   * Added OS metrics to the [metrics API](technical/web.md#modsmetrics).